Edit block content
Editing the text presents some challenges.

In short, we cannot retain the confidence/start/end values of
the return transcript if we dump a text into an edit box and let
the user just change things without constraint.

One key challenge is that we lose the start/end values (and any chance
of doing highligted word on play feature). I guess we could retain
the start/end values for the entire block?